Step B: TypedBuffer moves to the reference runtime; CL gets HostBuffer
D4, 2026-09-03: "put TypedBuffer in luminal_reference. CL shouldn't use
it for tests."

`TypedBuffer` and `ReferenceKernelCtx` were never core concepts. Core's
plans carry buffer IDs and opaque layouts, and nothing in the planner,
the bufferizer, or the IR types ever looked inside a payload — the
eleven-variant sum existed because the REFERENCE executor's kernels read
and write typed Rust slices. Both move to
`luminal_reference::typed_buffer` (552 lines out of
`src/buffer_tensor_ir.rs`), and `TypedBuffer` leaves core's prelude.
Core keeps the post-decision op surface, `BufferCopy`, `BufferAlloc`,
`BufferFree`.

CUDA-lite gets `host_buffer::HostBuffer` instead: bytes plus a dtype
tag, which is what a runtime whose kernels live on a device actually
needs. It makes the device bridge a memcpy — `typed_to_bytes` is
`&data.bytes` and `bytes_to_typed` is a constructor call, where both
used to be eleven-arm matches with five `unreachable!` arms. `From` for
`Vec<f32>`/`Vec<i32>`/`Vec<i64>`; Bool8 goes through the validated
`HostBuffer::bool8` constructor, NOT a `From<Vec<u8>>`, because a `From`
is by definition an unchecked door and Bool8 has exactly two legal
codes. `get_f32` returns `Vec<f32>` now (bytes cannot lend a `&Vec<T>`),
and `get_i32`/`get_i64`/`get_bool8` join it.

DIFFERENTIAL TESTS stage both payload types — the reference side a
`TypedBuffer`, the CL side a `HostBuffer`. That is the split working,
not a hole in it: `composed_read_families` names its data once as CL
payloads and carries a test-local `as_reference` bridge that decodes
them back through the dtype tag they were written under; the two
device-gated batteries build each side from the same `Vec<f32>`.
