+++ crates/luminal_cuda_lite/tests/cublaslt_contracts.rs
@@ -21,9 +21,9 @@
 #![cfg(feature = "device")]
 
 use cudarc::driver::{CudaContext, CudaSlice};
-use luminal::buffer_tensor_ir::TypedBuffer;
 use luminal::bufferize::BufferNode;
 use luminal::prelude::{DType, FxHashMap, NodeIndex};
+use luminal_cuda_lite::HostBuffer;
 
 /// The universal escape-and-disclose readback (the device_fidelity
 /// pattern): fetch the backing bytes + binding, walk each output
@@ -32,7 +32,7 @@ use luminal::prelude::{DType, FxHashMap, NodeIndex};
 fn walked_dense(rt: &CudaRuntime, out: NodeIndex) -> Vec<f32> {
     let (data, binding) = rt.fetch(out).expect("escape-and-disclose fetch");
     let bytes = match data {
-        TypedBuffer::F32(values) => values,
+        HostBuffer::F32(values) => values,
         other => panic!("output is {}, not f32", other.type_name()),
     };
     // The value's shape and read path both come from the RETURNED
@@ -355,16 +355,15 @@ fn marker_elected_bias_plan_matches_decomposed_route_tolerance_based() {
         let out = luminal_nn::linear(x, weight, Some(bias)).output();
         (cx, x.id, weight.id, bias.id, out.id)
     };
-    let data_for =
-        |x: NodeIndex, w: NodeIndex, b: NodeIndex| -> FxHashMap<NodeIndex, TypedBuffer> {
-            [
-                (x, TypedBuffer::from(weights(32, 1))),
-                (w, TypedBuffer::from(weights(24, 2))),
-                (b, TypedBuffer::from(weights(3, 3))),
-            ]
-            .into_iter()
-            .collect()
-        };
+    let data_for = |x: NodeIndex, w: NodeIndex, b: NodeIndex| -> FxHashMap<NodeIndex, HostBuffer> {
+        [
+            (x, HostBuffer::from(weights(32, 1))),
+            (w, HostBuffer::from(weights(24, 2))),
+            (b, HostBuffer::from(weights(3, 3))),
+        ]
+        .into_iter()
+        .collect()
+    };
     let options = luminal_cuda_lite::CompileOptions {
         generations: 12,
         generation_size: 16,
@@ -427,10 +426,10 @@ fn marker_elected_plan_matches_decomposed_route_tolerance_based() {
         let out = a.matmul(b).output();
         (cx, a, b, out)
     };
-    let data_for = |a: NodeIndex, b: NodeIndex| -> FxHashMap<NodeIndex, TypedBuffer> {
+    let data_for = |a: NodeIndex, b: NodeIndex| -> FxHashMap<NodeIndex, HostBuffer> {
         [
-            (a, TypedBuffer::from(weights(32, 1))),
-            (b, TypedBuffer::from(weights(24, 2))),
+            (a, HostBuffer::from(weights(32, 1))),
+            (b, HostBuffer::from(weights(24, 2))),
         ]
         .into_iter()
         .collect()
