+++ crates/luminal_reference/src/ops/cast/mod.rs
@@ -125,7 +125,7 @@ impl OpMatcher for CastMatcher {
 // 2026-08-13: everything about an op lives in the op's folder).
 // ---------------------------------------------------------------------------
 
-use luminal::buffer_tensor_ir::{ReferenceKernelCtx, TypedBuffer};
+use crate::typed_buffer::{ReferenceKernelCtx, TypedBuffer};
 
 pub(crate) fn kernel(
     _op: &dyn BufferTensorIrOp,
